refactor modal filters
